Move deployable_entity fixture to subdir
